Create new gpu_components.go to move vars
